+++ packages/web-core/CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Changed
 
 - Await for CLOSED WebSocket connection state after CLOSING one
+- Allow to change scene user parameters
 
 ## [2.7.1] - 2024-06-27